#260 - Fixed that the skip lines property in the Generator Application is an integer instead of a long.

+++ loa-service/loa-source-file-service/src/main/java/com/github/loa/source/file/configuration/FileDocumentSourceConfigurationProperties.java
@@ -31,5 +31,5 @@ public class FileDocumentSourceConfigurationProperties {
     /**
      * How many lines should be skipped before starting to process locations again.
      */
-    private final int skipLines;
+    private final long skipLines;
 }
